+++ public/mockServiceWorker.js
@@ -1,4 +1,3 @@
- 
 /* tslint:disable */
 
 /**
