--- src/features/this-must-be-deleted/Footer.stories.tsx
@@ -1,25 +1,0 @@
-import { Container } from "@/shared/components"
-import type { Meta, StoryObj } from "@storybook/react-vite"
-import Footer from "./Footer"
-
-const meta = {
-  title: "Footer",
-  component: Footer,
-  parameters: {
-    layout: "centered",
-  },
-  tags: ["autodocs"],
-} satisfies Meta<typeof Footer>
-
-export default meta
-
-type Story = StoryObj<typeof meta>
-
-export const Default: Story = {
-  args: {},
-  render: () => (
-    <Container width="xl">
-      <Footer />
-    </Container>
-  ),
-}
